feature(tiles.sass): ✨ Allow custom tile denominator
The default tile denominator is 12, but that can be a limiting factor. The ability to set it to 24 (or any other value) would allow us to use Bulma for front-end projects where a designer wouldn't be limited to 12 tiles (without having to split tiles in the same row into multiple groups of 12).